refactor: move tenant components from components/ to modules/tenants/components/
