+++ modules/tenants/components/footer/index.tsx
